// src/apis/models/FireBytes_Unified_Wave_ControlPanels_Dto_CreateGraphicPanelDevicePointViewInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_ControlPanels_Dto_CreateGraphicPanelDevicePointViewInput = {
    latitude?: number;
    longitude?: number;
    pointXvalue?: number;
    pointYvalue?: number;
    pointZvalue?: number;
    deviceId?: string | null;
    graphicPanelId?: string;
    isDefultBoundBuilding?: boolean;
    boundBuildingId?: string | null;
};